tiramisoon

+++ Makefile
@@ -5,10 +5,10 @@ SRC		:=	tiramisu.c callbacks.c
 CFLAGS	=	-Wall
 LDFLAGS	=	$(shell pkg-config --libs --cflags glib-2.0 gio-2.0)
 
-all: $(TARGET)
+all: clean $(TARGET)
 
 $(TARGET): $(OBJ)
 	$(CC) $(CFLAGS) $(LDFLAGS) $(SRC) -o $(TARGET)
 
 clean:
-	$(RM) ./tiramisu
+	@[ -f ./tiramisu ] && $(RM) ./tiramisu
